add coordinated activity managed signaling of the FSM.
If Start() returns a value it is serialized along with the activityID and sent in an FSM.ActivityStarted signal
If Tick() returns true,<non-nil value>,nil the non-nil value is serialized along with the activityID and sent in an FSM.ActivityUpdated signal

the update signals are considered to be important, so if the signal fails, we will fail the task and call Cancel() on the handler.

when ActivityStarted or ActivityUpdated signals are sent to the FSM, it recognizes them and extracts the activityID so activity related calls in the EventCorrelator will work too.