Introduce TestListener.Expects* to replace TestListener.Await*
This introduces a new set of methods to the TestLister that expects a
predicate to evaluate true after some action is taken. The plan is for
this to replace the Await* methods which are prone to race conditions if
the action takes longer than it takes to start awaiting. The
Register*Interest methods can also be made private after this change as
these are handled internally in a way that should avoid race conditions.
These additinal changes can be made at a later time after clients have
converted to this new faciclity.